parsers done only need to include bcg files, working on interactive map now.

+++ Strategic/insert.js
@@ -9,7 +9,7 @@ const BlendingPlantsParser = require('./BlendingPlantsParser.js');
 const SalesParser = require('./SalesParser.js');
 const PotentialAbujaParser = require('./PotentialAbujaParser.js');
 const FertlizerWeekPremiumParser = require('./PremiumWeekParser.js');
-
+const productionCapacitiesParser = require('./ProductionCapacitiesParser.js');
 
 const read = readline.createInterface({ input: process.stdin, output: process.stdout });
 
@@ -20,7 +20,9 @@ const jsonNames = ['BlendingPlantsData.json',
 	'Crudata.json',
 	'FertlizerWeekPremiumData.json',
 	'Potential-consumption-Data.json',
-	'SalesData.json'];
+	'SalesData.json',
+	'phosphatePlantProd.json',
+	'ureaPlantProd.json',];
 
 // const filesCode = [
 // 	'blendingPlants',
@@ -33,9 +35,11 @@ const fileNames = ['Blending plants Africa[1].xlsx',
 	'CRU FERTILIZER WEEK-Historical Prices Averages-Weekly Report (20230707)-60431[1].xlsx',
 	'Fertilizer Week Premium - report[1].xlsx',
 	'Potential consumption (Abuja + Agronomy)[1].xlsx',
-	'Sales 13-23[1].xlsx'];
+	'Sales 13-23[1].xlsx',
+	'Prod_capacities_Africa_23.xlsx'
+];
 
-const parsers = [BlendingPlantsParser, CruParser, FertlizerWeekPremiumParser, PotentialAbujaParser, SalesParser];
+const parsers = [BlendingPlantsParser, CruParser, FertlizerWeekPremiumParser, PotentialAbujaParser, SalesParser, productionCapacitiesParser];
 
 
 
@@ -103,10 +107,21 @@ async function insert(filename) {
 			console.log('parser returned undefined');
 			return;
 		};
-		const jsonFileName = jsonNames[fileNames.indexOf(filename)];
-		await writingJson(jsonizedPath, data, jsonFileName);
+		if(fileName !== "Prod_capacities_Africa_23.xlsx")
+		{
+			const jsonFileName = jsonNames[fileNames.indexOf(filename)];
+			await writingJson(jsonizedPath, data, jsonFileName);
+		}
+		else
+		{
+			for(let i = 0; i < data.length; i++)
+			{
+				const jsonFileName = jsonNames[fileNames.indexOf(filename) + i];
+				await writingJson(jsonizedPath, data[i], jsonFileName);
+			}
+		}
 	} catch (err) {
-		console.err('failed')
+		console.error('failed')
 	}
 	
 }
@@ -116,11 +131,12 @@ async function insert(filename) {
 async function run() {
 
 	const files = getFilesInDirectory(pathMaker('..', 'Strategic', 'xlsx'));
-	// console.log(files);
-	// return
-	for (const file of files) {
-		await insert(file);
-		deleteFile(pathMaker('..', 'Strategic', 'xlsx', file))
+	if(files.length !== 0)
+	{
+		for (const file of files) {
+			await insert(file);
+			deleteFile(pathMaker('..', 'Strategic', 'xlsx', file))
+		}
 	}
 	read.close();
